Update PostController.php
update IF in action

+++ backend/controllers/PostController.php
@@ -46,6 +46,8 @@ class PostController extends Controller
      */
     public function actionIndex()
     {
+        if(Yii::$app->user->can('post-list')) {
+
         $searchModel = new PostSerach();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -53,6 +55,8 @@ class PostController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+
+        }
     }
 
     /**
@@ -63,9 +67,11 @@ class PostController extends Controller
      */
     public function actionView($id)
     {
+        if(Yii::$app->user->can('post-view')) {
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
+        }
     }
 
     /**
@@ -75,6 +81,7 @@ class PostController extends Controller
      */
     public function actionCreate()
     {
+        if(Yii::$app->user->can('post-create')) {
         $model = new Post();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -84,6 +91,7 @@ class PostController extends Controller
         return $this->render('create', [
             'model' => $model,
         ]);
+        }
     }
 
     /**
@@ -95,6 +103,7 @@ class PostController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(Yii::$app->user->can('post-update')) {
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -104,6 +113,7 @@ class PostController extends Controller
         return $this->render('update', [
             'model' => $model,
         ]);
+        }
     }
 
     /**
@@ -115,9 +125,11 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
+        if(Yii::$app->user->can('post-delete')) {
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+        }
     }
 
     /**
